DataServices-api/src/test/java/eu/lifewatch/test/core/PairTest.java
package eu.lifewatch.test.core;

import eu.lifewatch.core.model.Pair;
import junit.framework.TestCase;
import org.junit.Test;

/**
 * @author Nikos Minadakis (minadakn 'at' ics 'dot' forth 'dot' gr)
 * @author Yannis Marketakis (marketak 'at' ics 'dot' forth 'dot' gr)
 */
public class PairTest extends TestCase{
    public PairTest(){
        super(PairTest.class.getName());
    }
    
    @Test
    public void testPlainCreation1(){
        String key="pair-key";
        String value="pair-value";
//      //  Pair pair=new Pair();
//        pair.setKey(key);
//        pair.setValue(value);
//        assertNotNull(pair);
//        assertEquals(key, pair.getKey());
//        assertEquals(value, pair.getValue());
        
        assert(true);
    }
    
    @Test
    public void testPlainCreation2(){
        String key="pair-key";
        String value="pair-value";
        Pair pair=new Pair(key,value);
        assertNotNull(pair);
        assertEquals(key, pair.getKey());
        assertEquals(value, pair.getValue());
    }
    
    @Test
    public void testStaticCreation(){
//        String key="pair-key";
//        String value="pair-value";
//        Pair pair=Pair.of(key,value);
//        assertNotNull(pair);
//        assertEquals(key, pair.getKey());
//        assertEquals(value, pair.getValue());
          assert(true);
    }
    
    @Test
    public void testComparePairs(){
//        String key="pair-key";
//        String value="pair-value";
//        Pair pair1=Pair.of(key, value);
//        /*Different key means different pairs*/
//        Pair pair2=Pair.of(key+"diff", value);
//        assertTrue(!pair1.equals(pair2));       
//        /*Different value means different pairs*/
//        pair2=Pair.of(key, value+"diff");
//        assertTrue(!pair1.equals(pair2));
//        /*Different key and value means different pairs*/
//        pair2=Pair.of(key+"diff", value+"diff");
//        assertTrue(!pair1.equals(pair2));
//        /*same key and value means same pairs*/
//        pair2=Pair.of(key, value);
//        assertTrue(pair1.equals(pair2));
//        assertEquals(pair1,pair2);
          assert(true);
    }
    
}